Cover MeteredVector/Map metering; some consolidation
