+++ soroban-env-host/benches/common/cost_types/mod.rs
@@ -4,11 +4,8 @@ mod compute_sha256_hash;
 mod ed25519_scalar_mul;
 mod guard_frame;
 mod host_obj_alloc_slot;
-mod im_map_immut_entry;
-mod im_map_mut_entry;
-mod im_vec_immut_entry;
-mod im_vec_mut_entry;
-mod im_vec_new;
+mod im_map_ops;
+mod im_vec_ops;
 mod record_contract_event;
 mod record_debug_event;
 mod scmap_to_host_map;
@@ -27,11 +24,8 @@ pub(crate) use compute_sha256_hash::*;
 pub(crate) use ed25519_scalar_mul::*;
 pub(crate) use guard_frame::*;
 pub(crate) use host_obj_alloc_slot::*;
-pub(crate) use im_map_immut_entry::*;
-pub(crate) use im_map_mut_entry::*;
-pub(crate) use im_vec_immut_entry::*;
-pub(crate) use im_vec_mut_entry::*;
-pub(crate) use im_vec_new::*;
+pub(crate) use im_map_ops::*;
+pub(crate) use im_vec_ops::*;
 pub(crate) use record_contract_event::*;
 pub(crate) use record_debug_event::*;
 pub(crate) use scmap_to_host_map::*;
