+++ soroban-env-host/src/cost_runner/cost_types/mod.rs
@@ -4,11 +4,8 @@ mod compute_sha256_hash;
 mod ed25519_scalar_mul;
 mod guard_frame;
 mod host_obj_alloc_slot;
-mod im_map_immut_entry;
-mod im_map_mut_entry;
-mod im_vec_immut_entry;
-mod im_vec_mut_entry;
-mod im_vec_new;
+mod im_map_ops;
+mod im_vec_ops;
 mod record_contract_event;
 mod record_debug_event;
 mod scmap_to_host_map;
@@ -29,11 +26,8 @@ pub use compute_sha256_hash::*;
 pub use ed25519_scalar_mul::*;
 pub use guard_frame::*;
 pub use host_obj_alloc_slot::*;
-pub use im_map_immut_entry::*;
-pub use im_map_mut_entry::*;
-pub use im_vec_immut_entry::*;
-pub use im_vec_mut_entry::*;
-pub use im_vec_new::*;
+pub use im_map_ops::*;
+pub use im_vec_ops::*;
 pub use record_contract_event::*;
 pub use record_debug_event::*;
 pub use scmap_to_host_map::*;
